refactor(session): remove SYNERGY_TEST_HOME fork from AgentTurn.stream
Replace the env-selected in-process stream fork with an explicit test-only
hook (AgentTurn.setInProcessStream) registered in test/preload.ts, moving
the fork body verbatim into src/session/agent-turn/in-process.ts and
sharing startContextUsageDraft via context-usage-draft.ts. Delete
AgentTurn.collectText (zero callers). Pool-path tests unregister the hook
via setInProcessStream(undefined) instead of deleting SYNERGY_TEST_HOME.
